+++ lib/json_schema_utils.dart
@@ -10,3 +10,4 @@ export 'src/boolean_json_schema.dart' show BooleanJsonSchema;
 export 'src/number_json_schema.dart' show NumberJsonSchema;
 export 'src/null_json_schema.dart' show NullJsonSchema;
 export 'src/object_json_schema.dart' show ObjectJsonSchema;
+export 'src/array_json_schema.dart' show ArrayJsonSchema;